feat bond layout
