+++ packages/accounts-app/src/Accounts/Staking/index.tsx
@@ -1,0 +1,5 @@
+// Copyright 2018-2019 @paritytech/substrate-light-ui authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+
+export * from './StakingOptions';
